feat: Implement binary genotype-phenotype model
This commit adds the implementation of a binary genotype-phenotype model in Python. The model is based on a mathematical framework proposed in a research paper by Yeonwoo Park. The implementation includes functions to calculate zeroth, first, and second-order effects, as well as to predict the phenotype value for a given genotype. The code has been tested with a simple binary genotype system with 3 positions and will be further tested with a 5-genotype system. This implementation aims to provide a reference-free analysis of genotype-phenotype relationships.

[END COMMIT MESSAGE]