refactor: Generalize code and add new functions
Refactor the code to generalize the conditions and add new functions
for calculating genotype averages and creating genotypes.
This includes the addition of the `all_sites_match_states` function,
the `make_comprehensive_genotypes` function, and the `get_indices_with_genotypes` function.
These changes aim to improve the code's flexibility and maintainability.